+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* Generation of C/C++ bindings added in the Python interface and included in the test suite
 * Support in Rust for Cartesian product of constraints
 
 ### Removed